requirements.txt
pytorch==1.9.0
torchvision==0.10.0
cudatoolkit==11.1
pytorch-lightning==1.6.1
torchnet==0.0.4
opencv-python==4.7
opencv-contrib-python==4.5.5
